requirements.txt
alabaster==0.7.12
attrs==21.4.0
Babel==2.9.1
beautifulsoup4==4.11.1
bleach==5.0.1
Bottleneck==1.3.5
brotlipy==0.7.0
certifi==2022.9.24
cffi==1.15.0
charset-normalizer==2.0.4
colorama==0.4.4
commonmark==0.9.1
cryptography==37.0.1
cycler==0.11.0
defusedxml==0.7.1
docutils==0.15.2
entrypoints==0.4
fastjsonschema==2.16.2
fonttools==4.25.0
future==0.18.2
idna==3.3
imagesize==1.3.0
importlib-metadata==4.11.3
importlib-resources==5.10.0
iniconfig==1.1.1
Jinja2==3.0.3
joblib==1.1.1
jsonschema==4.16.0
jupyter_client==7.4.4
jupyter_core==4.11.1
jupyterlab-pygments==0.2.2
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.2
mistune==2.0.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
munkres==1.1.4
nbclient==0.7.0
nbconvert==7.2.3
nbformat==5.7.0
nbsphinx==0.8.9
nest-asyncio==1.5.6
numexpr==2.8.3
numpy
packaging==21.3
pandas==1.4.4
pandocfilters==1.5.0
Pillow==9.2.0
pip==22.2.2
pkgutil_resolve_name==1.3.10
pluggy==1.0.0
py==1.11.0
pycausalgps==0.0.1
pycparser==2.21
Pygments==2.11.2
pyOpenSSL==22.0.0
pyparsing==3.0.4
pyrsistent==0.18.1
PySocks==1.7.1
pytest==7.1.2
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
pyzmq==24.0.1
recommonmark==0.6.0
requests==2.27.1
scikit-learn==1.1.2
scipy==1.9.1
setuptools==65.5.0
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
Sphinx==4.4.0
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlitedict==2.0.0
threadpoolctl==2.2.0
tinycss2==1.2.1
tomli==2.0.1
tornado==6.2
traitlets==5.5.0
typing_extensions==4.4.0
urllib3==1.26.9
webencodings==0.5.1
wheel==0.37.1
xgboost==1.5.0
zipp==3.8.0
